Bugfix/myw 1063 shopping points amount type fix (#195)
* MYW-856 moved MyWorld PSP payment confirmation and validity check to order post save plugin

* Removed spy_price_product_store constraint until it's clear why we need it.

* myw-856 added MyWorld payment saving in the database with a relation to a sales order

* myw-856 save myworld payment data for unsuccessful responses as well

* myw-856 moved payment response saving before handling api response data

* myw-1063 hotfix for shopping points calculator (int => float)

+++ src/Pyz/Zed/MyWorldPayment/Business/Calculator/ShoppingPointsPaymentCalculator.php
@@ -52,13 +52,13 @@ class ShoppingPointsPaymentCalculator implements MyWorldPaymentCalculatorInterfa
     /**
      * @param \Generated\Shared\Transfer\CalculableObjectTransfer $calculableObjectTransfer
      *
-     * @return int
+     * @return float
      */
-    protected function getCommonUsedShoppingPoints(CalculableObjectTransfer $calculableObjectTransfer): int
+    protected function getCommonUsedShoppingPoints(CalculableObjectTransfer $calculableObjectTransfer): float
     {
         return array_reduce(
             $calculableObjectTransfer->getItems()->getArrayCopy(),
-            function (int $carry, ItemTransfer $itemTransfer) {
+            function (float $carry, ItemTransfer $itemTransfer) {
                 $carry += $itemTransfer->getTotalUsedShoppingPointsAmount();
 
                 return $carry;
